+++ src/SvgIcons.h
@@ -30,6 +30,7 @@ enum class TbIcon {
     Close,
     ArrowsDiagonal,
     ArrowsDiagonalMinimize,
+    MatchWholeWord,
     kMax
 };